# contratos/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "auto_cusco API",
    "description": "API REST del backend de auto_cusco: ingesta de sabanas, seleccion y segmentacion de cartera, generacion de cargas (digitales y VoIP) y reportes de cobranza.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/cargas": {
      "post": {
        "tags": [
          "cargas"
        ],
        "summary": "Subir Sabana",
        "description": "Registra una version nueva y lanza su procesamiento en segundo plano (V-1).",
        "operationId": "subir_sabana_cargas_post",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_subir_sabana_cargas_post"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CargaCreadaRespuesta"
                }
              }
            }
          },
          "400": {
            "description": "La peticion no se puede atender tal como viene; el motivo esta en detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetalleError"
                }
              }
            }
          },
          "413": {
            "description": "El archivo supera el tamano maximo",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetalleError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "cargas"
        ],
        "summary": "Listar Versiones",
        "operationId": "listar_versiones_cargas_get",
        "parameters": [
          {
            "name": "fecha_corte",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Fecha Corte"
            }
          },
          {
            "name": "limite",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limite"
            }
          },
          {
            "name": "desplazamiento",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Desplazamiento"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VersionRespuesta"
                  },
                  "title": "Response Listar Versiones Cargas Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cargas/{carga_id}": {
      "get": {
        "tags": [
          "cargas"
        ],
        "summary": "Obtener Version",
        "operationId": "obtener_version_cargas__carga_id__get",
        "parameters": [
          {
            "name": "carga_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Carga Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VersionRespuesta"
                }
              }
            }
          },
          "404": {
            "description": "No existe lo pedido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetalleError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "cargas"
        ],
        "summary": "Eliminar Version",
        "operationId": "eliminar_version_cargas__carga_id__delete",
        "parameters": [
          {
            "name": "carga_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Carga Id"
            }
          },
          {
            "name": "dejar_fecha_sin_vigente",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Confirma que la fecha quede sin version vigente (V-8)",
              "default": false,
              "title": "Dejar Fecha Sin Vigente"
            },
            "description": "Confirma que la fecha quede sin version vigente (V-8)"
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "description": "No existe lo pedido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetalleError"
                }
              }
            }
          },
          "409": {
            "description": "La operacion choca con el estado actual",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetalleError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/cartera": {
      "get": {
        "tags": [
          "cartera"
        ],
        "summary": "Consultar Cartera",
        "description": "Productos de la version vigente de la fecha (RF-04, RF-05, RF-08).",
        "operationId": "consultar_cartera_cartera_get",
        "parameters": [
          {
            "name": "fecha_corte",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Fecha Corte"
            }
          },
          {
            "name": "filtro",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "default": [],
              "title": "Filtro"
            }
          },
          {
            "name": "orden",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "campo, o -campo para descendente",
              "title": "Orden"
            },
            "description": "campo, o -campo para descendente"
          },
          {
            "name": "limite",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 5000,
              "minimum": 1,
              "default": 50,
              "title": "Limite"
            }
          },
          {
            "name": "desplazamiento",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Desplazamiento"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginaRespuesta"
                }
              }
            }
          },
          "400": {
            "description": "La peticion no se puede atender tal como viene; el motivo esta en detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetalleError"
                }
              }
            }
          },
          "404": {
            "description": "No existe lo pedido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetalleError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/archivos-carga/previsualizacion": {
      "post": {
        "tags": [
          "archivos de carga"
        ],
        "summary": "Previsualizar",
        "description": "Primeras filas y errores de la carga, sin escribir el archivo.\n\n`disponibles` es el total de la seleccion completa, no el de la muestra: es\nlo que dice si alcanzan los productos para la cantidad pedida (RF-08).",
        "operationId": "previsualizar_archivos_carga_previsualizacion_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PeticionGeneracion"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrevisualizacionRespuesta"
                }
              }
            }
          },
          "400": {
            "description": "La peticion no se puede atender tal como viene; el motivo esta en detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetalleError"
                }
              }
            }
          },
          "404": {
            "description": "No existe lo pedido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetalleError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/archivos-carga": {
      "post": {
        "tags": [
          "archivos de carga"
        ],
        "summary": "Generar Archivo",
        "description": "Genera la carga completa y devuelve el archivo (RF-09, RF-13, RF-15).\n\nEl resumen viaja en cabeceras `X-Carga-*` porque el cuerpo es el archivo;\nel detalle de los errores se consulta en `/previsualizacion`.",
        "operationId": "generar_archivo_archivos_carga_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PeticionGeneracion"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "El archivo de carga en el formato pedido",
            "headers": {
              "Content-Disposition": {
                "description": "attachment; filename=\"<nombre>.<extension>\"",
                "schema": {
                  "type": "string"
                }
              },
              "X-Carga-Generados": {
                "description": "Filas escritas en el archivo",
                "schema": {
                  "type": "string",
                  "pattern": "^[0-9]+$"
                }
              },
              "X-Carga-Disponibles": {
                "description": "Productos que cumplen la seleccion completa",
                "schema": {
                  "type": "string",
                  "pattern": "^[0-9]+$"
                }
              },
              "X-Carga-Solicitados": {
                "description": "Cantidad de productos pedida",
                "schema": {
                  "type": "string",
                  "pattern": "^[0-9]+$"
                }
              },
              "X-Carga-Suficiente": {
                "description": "1 si habia al menos tantos productos como se pidieron, 0 si no",
                "schema": {
                  "type": "string",
                  "enum": [
                    "0",
                    "1"
                  ]
                }
              },
              "X-Carga-Errores": {
                "description": "Campos que no se pudieron generar; el detalle esta en /previsualizacion",
                "schema": {
                  "type": "string",
                  "pattern": "^[0-9]+$"
                }
              }
            },
            "content": {
              "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "text/csv": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "La peticion no se puede atender tal como viene; el motivo esta en detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetalleError"
                }
              }
            }
          },
          "404": {
            "description": "No existe lo pedido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetalleError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_subir_sabana_cargas_post": {
        "properties": {
          "fecha_corte": {
            "type": "string",
            "format": "date",
            "title": "Fecha Corte",
            "description": "Fecha del corte que representa la sabana"
          },
          "archivo": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Archivo",
            "description": "Archivo de la sabana (.xlsb, .xlsx)"
          },
          "hoja": {
            "type": "string",
            "title": "Hoja",
            "default": "VENCIDA"
          }
        },
        "type": "object",
        "required": [
          "fecha_corte",
          "archivo"
        ],
        "title": "Body_subir_sabana_cargas_post"
      },
...
      "CargaCreadaRespuesta": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "fecha_corte": {
            "type": "string",
            "format": "date",
            "title": "Fecha Corte"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "estado": {
            "$ref": "#/components/schemas/EstadoCarga"
          },
          "versiones_identicas": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Versiones Identicas"
          },
          "aviso": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Aviso"
          }
        },
        "type": "object",
        "required": [
          "id",
          "fecha_corte",
          "version",
          "estado",
          "versiones_identicas",
          "aviso"
        ],
        "title": "CargaCreadaRespuesta"
      },
...
      "DetalleError": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "DetalleError",
        "description": "Motivo legible por el que la API rechazo la peticion."
      },
...
      "EstadoCarga": {
        "type": "string",
        "enum": [
          "en_cola",
          "procesando",
          "terminada",
          "fallida"
        ],
        "title": "EstadoCarga",
        "description": "Estado de procesamiento de una version (la ingesta corre en segundo plano)."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaginaRespuesta": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limite": {
            "type": "integer",
            "title": "Limite"
          },
          "desplazamiento": {
            "type": "integer",
            "title": "Desplazamiento"
          },
          "suficiente": {
            "type": "boolean",
            "title": "Suficiente"
          },
          "productos": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Productos"
          }
        },
        "type": "object",
        "required": [
          "total",
          "limite",
          "desplazamiento",
          "suficiente",
          "productos"
        ],
        "title": "PaginaRespuesta"
      },
...
      "PeticionGeneracion": {
        "properties": {
          "fecha_corte": {
            "type": "string",
            "format": "date",
            "title": "Fecha Corte"
          },
          "definicion": {
            "$ref": "#/components/schemas/DefinicionEntrada"
          },
          "formato": {
            "$ref": "#/components/schemas/FormatoArchivo",
            "default": "xlsx"
          },
          "filtros": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Filtros",
            "default": []
          },
          "orden": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Orden"
          },
          "cantidad": {
            "type": "integer",
            "maximum": 120000.0,
            "minimum": 1.0,
            "title": "Cantidad",
            "default": 1000
          },
          "opciones": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OpcionesEntrada"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "fecha_corte",
          "definicion"
        ],
        "title": "PeticionGeneracion"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VersionRespuesta": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "fecha_corte": {
            "type": "string",
            "format": "date",
            "title": "Fecha Corte"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "estado": {
            "$ref": "#/components/schemas/EstadoCarga"
          },
          "vigente": {
            "type": "boolean",
            "title": "Vigente"
          },
          "nombre_archivo": {
            "type": "string",
            "title": "Nombre Archivo"
          },
          "tamano_bytes": {
            "type": "integer",
            "title": "Tamano Bytes"
          },
          "hoja": {
            "type": "string",
            "title": "Hoja"
          },
          "creado_en": {
            "type": "string",
            "format": "date-time",
            "title": "Creado En"
          },
          "procesado_en": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Procesado En"
          },
          "huella_formato": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Huella Formato"
          },
          "filas_total": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filas Total"
          },
          "filas_ingestadas": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filas Ingestadas"
          },
          "incidencias_error": {
            "type": "integer",
            "title": "Incidencias Error",
            "default": 0
          },
          "incidencias_advertencia": {
            "type": "integer",
            "title": "Incidencias Advertencia",
            "default": 0
          },
          "incidencias_info": {
            "type": "integer",
            "title": "Incidencias Info",
            "default": 0
          },
          "motivo_fallo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Motivo Fallo"
          }
        },
        "type": "object",
        "required": [
          "id",
          "fecha_corte",
          "version",
          "estado",
          "vigente",
          "nombre_archivo",
          "tamano_bytes",
          "hoja",
          "creado_en",
          "procesado_en",
          "huella_formato",
          "filas_total",
          "filas_ingestadas",
          "incidencias_error",
          "incidencias_advertencia",
          "incidencias_info",
          "motivo_fallo"
        ],
        "title": "VersionRespuesta"
      }
    }
  }
}